// apps/web/data/unique-data.generated.ts
export interface UniqueItemData {
  name: string;
  base: string;
  slot: string;
  mods: string[];
}